requirements.txt
python-telegram-bot==13.15
requests==2.31.*
environs==9.5.*